Rename DiscoUtil implementation
